% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Data-Cleansing.R
\docType{methods}
\name{accederFilas}
\alias{accederFilas}
\alias{filaError,Fila-method}
\alias{filaError<-,Fila-method}
\alias{filaColumnas,Fila-method}
\alias{filaColumnas<-,Fila-method}
\alias{filaColNombre,Fila-method}
\alias{filaColTipo,Fila-method}
\alias{filaMensajeError,Fila-method}
\alias{filaMensajeError<-,Fila-method}
\alias{filaColumnaEnError,Fila-method}
\alias{filaColumnaEnError<-,Fila-method}
\alias{filaStatus,Fila-method}
\alias{filaStatus<-,Fila-method}
\title{Configura el valor de los elementos de \linkS4class{Fila}.}
\usage{
\S4method{filaError}{Fila}(x)

\S4method{filaError}{Fila}(x) <- value

\S4method{filaColumnas}{Fila}(x)

\S4method{filaColumnas}{Fila}(x) <- value

\S4method{filaColNombre}{Fila}(x)

\S4method{filaColTipo}{Fila}(x)

\S4method{filaMensajeError}{Fila}(x)

\S4method{filaMensajeError}{Fila}(x) <- value

\S4method{filaColumnaEnError}{Fila}(x)

\S4method{filaColumnaEnError}{Fila}(x) <- value

\S4method{filaStatus}{Fila}(x)

\S4method{filaStatus}{Fila}(x) <- value
}
\description{
Estas funciones pueden usarse para que usuarios del paquete puedan crear el método de \link{ValidarFila} más fácilmente. Ver ejemplo en \link{ValidarFIla}.
}
\section{Functions}{
\itemize{
\item \code{filaError,Fila-method}: Accede un mensaje de error en una fila

\item \code{filaError<-,Fila-method}: Guarda un mensaje de error en una fila

\item \code{filaColumnas,Fila-method}: Accede las columnas una fila

\item \code{filaColumnas<-,Fila-method}: Guarda las columnas una fila

\item \code{filaColNombre,Fila-method}: Accede el nombre de los campos en una fila.

\item \code{filaColTipo,Fila-method}: Accede el tipo de dato de los campos en una fila.

\item \code{filaMensajeError,Fila-method}: Accede un mensaje de error

\item \code{filaMensajeError<-,Fila-method}: \bold{Anexa} un mensaje de error

\item \code{filaColumnaEnError,Fila-method}: Accede la posición de la columna en error

\item \code{filaColumnaEnError<-,Fila-method}: \bold{Anexa} la posición de la columna en error

\item \code{filaStatus,Fila-method}: Accede el estatus de la fila

\item \code{filaStatus<-,Fila-method}: Guarda el estatus de la fila
}}

