% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Data-Cleansing.R
\name{AgregarFila}
\alias{AgregarFila}
\title{Agrega una fila a la colección.}
\usage{
AgregarFila(Registro, Fila, Filas, contador)
}
\arguments{
\item{Registro}{es una fila del data frame. Las columnas deben tener los mismos nombres que el slot del objeto \code{Fila@Columnas}.}

\item{Fila}{Un objeto \linkS4class{Fila} o descendiente.}

\item{Filas}{un objeto \linkS4class{Filas} donde el objeto \code{Fila} será guardado.}
}
\value{
\linkS4class{Filas} o descendiente.
}
\description{
Agrega una \linkS4class(Fila) a la colección de \linkS4class{Filas}. Antes de agregar la fila a la colección, llama a la función genérica \link{ValidarFila}; si encuentra un método apropiado, lo despacha y luego agrega la fila a la colección; de lo contrario, solo agrega la fila a la colección sin realizar validación. En el primer caso, el \emph{slot} \code{FueValidado = TRUE} y en el segundo caso su valor es \code{FALSE}.
}
