% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Data-Cleansing.R
\docType{class}
\name{Columnas-class}
\alias{Columnas-class}
\alias{Columnas}
\title{Clase Columnas}
\usage{
Columnas(Nombres, Tipos)
}
\description{
Un objeto S4 para modelar una columna.
}
\details{
No pertenece a un sub-sistema directamente. Es una clase de apoyo.

Contiene los nombres y los tipos de datos de una tabla.
}
\section{Functions}{
\itemize{
\item \code{Columnas}: Constructor de columnas.
}}

\section{Slots}{

\describe{
\item{\code{Nombres}}{un vector de texto con el nombre de las columnas.}

\item{\code{Tipos}}{un vector de texto con los tipos de datos de las columnas.}
}}

