% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-Extract.R
\docType{class}
\name{Conexion-class}
\alias{Conexion-class}
\title{Clase Conexión}
\description{
Un objeto S4 para gestionar conexiones a archivos de datos o bases de datos.
}
\details{
Pertenece al \bold{Sub-sistema 3 de extracción.}

Maneja información específica de la conexión, que luego permite a interfaces comunes, conectarse con el archivo apropiado para extraer o almacenar datos.

Actualmente, soporta conexiones a RDS, SQLite. Para construir la conexión usar la función \code{Conexion<ARCHIVO>}. Por ejemplo: \code{ConexionCSV(...)}, \code{ConexionSQLite(...)}.
}
\section{Slots}{

\describe{
\item{\code{connString}}{Una cadena de texto con las instrucciones para realizar la conexión.}
}}

