% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-Extract.R
\docType{class}
\name{ConexionRDS-class}
\alias{ConexionRDS-class}
\title{Clase Conexión RDS}
\description{
Un objeto S4 para gestionar conexiones objetos de datos generados por R.
}
\details{
Pertenece al \bold{Sub-sistema 3 de extracción.}
}
