% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-Extract.R
\docType{class}
\name{ConexionSQLite-Class}
\alias{ConexionSQLite-Class}
\title{Clase Conexión SQLite}
\description{
Un objeto S4 para gestionar conexiones a una base de datos en SQLite, a través del paquete \link[=package]{RSQLite}.
}
\section{Slots}{

\describe{
\item{\code{Driver}}{El driver para la conexión con SQLite.
Al usar \link{ConexionSQLite} para crear la conexión, el driver se asigna automáticamente.}

\item{\code{Tabla}}{El nombre de la tabla donde se desea guardar los datos.}
}}

