% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-Extract.R
\name{ConexionSQLite}
\alias{ConexionSQLite}
\title{ConexionSQLite}
\usage{
ConexionSQLite(RutaDB, Tabla)
}
\arguments{
\item{RutaDB}{La ruta al archivo con la base de datos.}

\item{Tabla}{El nombre de la tabla en la base de datos.}
}
\description{
Crea una conexión a un archivo de SQLite.
}
\examples{

\dontrun{
   ConexionSQLite("C:/myDB.sqlite") # Añadir extensión, o arroja error
    ConexionSQLite("C:/myDB.db") # Cambia extension
}
}
