% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Data-Cleansing.R
\docType{methods}
\name{extractoresFilas}
\alias{extractoresFilas}
\alias{filasEnError,Filas-method}
\alias{filasErrores,Filas-method}
\alias{filasErroresPos,Filas-method}
\alias{filasCorrectas,Filas-method}
\alias{filasStatus,Filas-method}
\title{Acceder a elementos de  \linkS4class{Filas}.}
\usage{
\S4method{filasEnError}{Filas}(Filas)

\S4method{filasErrores}{Filas}(Filas)

\S4method{filasErroresPos}{Filas}(Filas)

\S4method{filasCorrectas}{Filas}(Filas)

\S4method{filasStatus}{Filas}(Filas)
}
\description{
Acceder a elementos de  \linkS4class{Filas}.
}
\section{Functions}{
\itemize{
\item \code{filasEnError,Filas-method}: Extrae filas en error (vector lógico)

\item \code{filasErrores,Filas-method}: Extrae lista de errores

\item \code{filasErroresPos,Filas-method}: Extrae la posición de las filas con errores

\item \code{filasCorrectas,Filas-method}: Extrae las filas correctas (vector lógico)

\item \code{filasStatus,Filas-method}: Extrae el estatus de las filas
}}

