% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genericos.R, R/03-Extract-ConexionRDS.R,
%   R/03-Extract-ConexionRSQLite.R
\docType{methods}
\name{Extraer-methods}
\alias{Extraer-methods}
\alias{Extraer,ConexionRDS-method}
\alias{Extraer,ConexionSQLite-method}
\title{Extraer}
\usage{
Extraer(conexion, ...)

\S4method{Extraer}{ConexionRDS}(conexion)

\S4method{Extraer}{ConexionSQLite}(conexion)
}
\arguments{
\item{conexion}{Una conexión al archivo o base de datos donde se desea guardar el objeto.}
}
\value{
El tipo de objeto depende de la conexión.
}
\description{
Extrae un objeto, usando una conexión.

Extrae un objeto R de una conexión \link{ConexionRDS}.
}
\details{
Esta función requiere de la conexión, sin ningún otro detalle adicional.
}
\section{Functions}{
\itemize{
\item \code{Extraer,ConexionRDS-method}: Extrar un objeto R en en una conexión \linkS4class{ConexionRDS}.

\item \code{Extraer,ConexionSQLite-method}: Extrae un data.frame de una conexión a SQLite.
}}

\section{Condiciones y Errores}{


Devuelve error si el objeto no se encuentra en la ruta provista.
}

\examples{
\dontrun{
  objeto <- Extraer(miConexioRDS)
}

}
