% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genericos.R, R/03-Extract-ConexionRDS.R
\docType{methods}
\name{Extraer}
\alias{Extraer}
\alias{Extraer,ConexionRDS-method}
\title{Extraer}
\usage{
Extraer(conexion, ...)

\S4method{Extraer}{ConexionRDS}(conexion)
}
\arguments{
\item{conexion}{Una conexión al archivo o base de datos donde se desea guardar el objeto.}
}
\value{
El tipo de objeto depende de la conexión.
}
\description{
Extrae un objeto, usando una conexión.

Extrae un objeto R de una conexión \link{ConexionRDS}.
}
\details{
Esta función requiere de la conexión, sin ningún otro detalle adicional.
}
\section{Methods (by class)}{
\itemize{
\item \code{ConexionRDS}: Extrar un objeto R en en una conexión \linkS4class{ConexionRDS}.
}}

\section{Condiciones y Errores}{


Devuelve error si el objeto no se encuentra en la ruta provista.
}

\examples{
\dontrun{
  objecto <- Extraer(miConexioRDS)
}

}
