% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Data-Cleansing.R
\docType{class}
\name{Fila-class}
\alias{Fila-class}
\alias{Fila}
\title{Clase Fila}
\usage{
Fila(Columnas)
}
\description{
Un objeto S4 para modelar una Fila
}
\details{
Contiene los nombres y los tipos de datos de la fila de una una tabla. Además su principal función es dar seguimiento al estado de la fila; es decir, si contiene errores, está siendo actualizada, es nueva, etc...
}
\section{Functions}{
\itemize{
\item \code{Fila}: Constructor de fila
}}

\section{Slots}{

\describe{
\item{\code{Columnas}}{un objecto con las \linkS4class{Columnas} de una tabla.}

\item{\code{Error}}{indica si la fila contiene un error en algún valor de sus columnas.}

\item{\code{ErroresMsg}}{un vector de texto, con un elemento por columna en error.}

\item{\code{ErroresIndice}}{un vector de enteros con la(s) posición(es) de la(s) columna(s) en error.}

\item{\code{Status}}{el status de la columna. Una lista enumerada ('Nueva', 'Eliminada','Cambiada','Sin cambio')}
}}

\section{Uso previsto}{


\code{Fila} es un objeto auxiliar y no contiene datos en sí mismo. La función \link{TraerFilas(data, Filas)} se encarga de filtrar un \code{data.frame} extrayendo las filas correctas; mientras que, \code{TraerErrores(data, Filas)} extrae las filas erróneas en su propio \code{data.frame}.

La función de este objeto, es llevar el estado de una fila en procesos de verificación, y procesamiento de registros.
}

