% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Data-Cleansing.R
\docType{class}
\name{Filas-class}
\alias{Filas-class}
\title{Clase Filas}
\description{
Un objeto S4 para modelar varias Filas
}
\details{
No pertenece a un sub-sistema directamente. Es una clase de apoyo.

Contiene un resumen del estado de las filas de una tabla, pero no contiene sus datos.
}
\section{Slots}{

\describe{
\item{\code{Columnas}}{un objeto con las \linkS4class{Columnas} de una tabla.}

\item{\code{Error}}{indica si alguna de todas las filas contiene un error}

\item{\code{Errores}}{contiene la lista de errores. Cada elemento de la lista contiene los errores de una fila.}

\item{\code{FilasEnError}}{un vector numérico que identifica la posición de las filas con error.}

\item{\code{ColumnasEnError}}{una lista donde cada elemento contiene un vector númerico con el índice de la columna que contiene el error.}

\item{\code{NombreTabla}}{el nombre de la tabla.}

\item{\code{FilasCorrectas}}{un vector numérico que identifica la posición de las filas sin errores.}

\item{\code{Status}}{un vector con el estatus de cada fila.}
}}

