% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-Deduplication.R
\name{NormalizarTexto}
\alias{NormalizarTexto}
\title{Normalizar  texto}
\usage{
NormalizarTexto(text)
}
\arguments{
\item{text}{el texto que se desea normalizar.}
}
\value{
Una lista con el texto limpio (\code{text$Limpio}) y el texto original (\code{text$Sucio})
}
\description{
Tomar un texto, elimina los espacios, los signos de puntuación, y lo sube a mayúsculas. Usar para aumentar las probabilidades de eliminar duplicados por mala escritura, o como paso previo a un fuzzy matching con \link{EncontrarSurrogateFuzzyMemo}.
}
\details{
Pertenece al \bold{sub-sistema 07 de de-duplicación}.
}
