% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Data-Cleansing.R
\name{TraerErrores}
\alias{TraerErrores}
\title{Devuelve las filas que \bold{no pasan} la validaci贸n.}
\usage{
TraerErrores(Filas, Datos)
}
\arguments{
\item{Filas}{Un objeto \linkS4class{Filas}.

 @name Data-Extractors
 @rdname Data-Extractors}

\item{Datos}{un \link{data.frame}.}
}
\value{
un \link{data.frame}.
}
\description{
Requiere de una instancia de \linkS4class{Filas} generada por \link{ValidarFilas}, y un \code{data.frame}. La funci贸n mira entre las filas que contienen errores, y filtra el conjunto de datos para devolver solo las filas incorrectas.
}
\seealso{
la funci贸n \link{TraerFilas} que extrae las filas err贸nea
}
