% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Data-Cleansing.R
\name{TraerFilas}
\alias{TraerFilas}
\title{Devuelve las filas que pasan la validación}
\usage{
TraerFilas(Filas, Datos)
}
\arguments{
\item{Filas}{Un objeto \linkS4class{Filas}.}

\item{Datos}{un \link{data.frame}.}
}
\value{
un \link{data.frame}.
}
\description{
Requiere de una instancia de \linkS4class{Filas} generada por \link{ValidarFilas}, y un \code{data.frame}. La función mira entre las filas que han pasado la validación, y filtra el conjunto de datos para devolver solo las filas correctas.
}
\seealso{
la función \link{TraerErrores} que extrae las filas errónea

 @name Data-Extractors
 @rdname Data-Extractors
}
