% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Data-Cleansing.R
\name{ValidarFilas}
\alias{ValidarFilas}
\title{Valida todas las filas de un data frame.}
\usage{
ValidarFilas(Fila, Datos)
}
\arguments{
\item{Fila}{Un objeto \linkS4class{Fila} o que describe las filas de los \code{Datos}.}

\item{Datos}{un \link{data.frame}.}
}
\value{
\linkS4class{Filas} o descendiente.
}
\description{
Si existe un método para validar filas, (una implementación de la función genérica \link{ValidarFila}); lo utiliza para validar todas las filas de un data.frame. Devuelve una colección de \linkS4class{Filas}, que luego pueden usarse para obtener las filas limpias y las filas en error del \link{data.frame}.
}
\details{
Utilizar el objeto \code{Filas}, devuelto por esta función para usarlo en conjunto con la función \link{TraerFilas(data, Filas)} que se encarga de filtrar un \code{data.frame} extrayendo las filas correctas; y la función \code{TraerErrores(data, Filas)} que extrae las filas erróneas.
}
